Crate prac

a practice-cultivating utility

(for overachieving procrastinators recovering from an unhealthy fixation on self-development)

What this isn’t

This is not a todo list, a calendar, pomodoro timer, a scheduling app, or a habit tracker.

Design decisions have been made on ideological grounds to intentionally maximize incompatibility with these types of apps. Neither is this a compliance mechanism, if you wish to Pavlov yourself into a life well lived, that is your prerogative, but I will not be helping you.

What this is

A tool to augment a practice-driven life/workflow, specifically providing feedback to enable those practices which cannot all be done daily, or the efficacy of which is highly sensitive to factors knowable only in the moment.

In a state of immersion, time is experienced. In many productivity systems, time is controlled–better had than spent, better spent than lost.

That REM sleep lifts us gracefully through from one phase of deep sleep to the next at an average period of ~90 minutes does not make possible reproduction of the effect by way of a 90-minute alarm clock. While we cannot stay in flow forever, we can remain in the experience of time as we allow our natural stuckness to lead our transition to the next state of practice. This is the motivating philosophy of prac.

If you want, you can get started right now with prac help. If you want to see how I attempted to integrate the dream theory of flow state into, read on to the minifesto.

Minifesto and user guide

Begin by negatively scheduling, making time with guarantee only that you will not permit yourself to be otherwise scheduled or interrupted (including by phones/notifications).

After having prac add-ed a few practices, you can prac list and choose one to start on (how you choose is none of my business).

This feedback loop orients on starts rather than completions, banking on that (canonically) self-promising to make it through the gym doors will be more successful to motivate a workout than a list of the entailed exercises. Output, sharing, and interpersonal feedback are no less important than starts but happen to inhabit another segment of the loop, whereas this tool limits its scope to the “refocusing” stage.

In a conventional productivity system, interrupts are triggered externally, by calendar notifications, timers, due dates, etc. There are few ways in which these systems could be less accountable to the psychological state of the user in the relevant moment. The discipline of producing conventional productivity tools might as well be called “distraction engineering.”

Is genuine feedback even possible under a system where all the decisions are made ahead of time? When we are most efficient, it might not appear to us that we are making a decision at all, whether to continue or stop. Clearly then the problem is not implementation but orientation.

It’s not about when you work, not everything is going to get done, you have X hours no matter what (and should probably limit yourself to even fewer), but more about when you switch between tasks.

In a mode of practice, control is not exerted by the clock, but follows naturally from the persons instincts of relative flow and stuckness.

When you get stuck, rather than banging your head until your pomodoro takes pity on your soul, you simply prac list to see a handful of tasks with a progress bar showing how long it’s been since you last practiced as a fraction of how frequently you wish to practice. This provides a very gentle way prioritize those practices that have been recently neglected.

Looks like this:

prac list

distributed systems programming ▬▬▬
                      daily log ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
                       exercise ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
                    kierkegaard ▬▬▬▬▬▬▬
                          steno ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
                    weekly repo ▬▬▬

(tip: use `prac list --cumulative` to see cumulative hours logged, are we 10000 yet?)

In prac the clock has no control and appears only as purely geometric information to augment the human weakness that is ascertaining one’s position in the passage of time.

For those who struggle with work-life balance, these tasks can include things like rest, play, socializing, eating, outdoor time, family time, and other practices of self-care.

I firmly believe that time and quality of practice account for the bulk of competence, and so I’ve implemented only two tracking features, proc trac for a bare total of time, and prac log for plain text goal-setting and reflection.

It’s not a big deal if you let it run over, take it as a sign that you should extend your period. If you find yourself regularly finishing early, you’ve identified that you would benefit from a shorter feedback cycle!

Again, it’s not a routine app, but I think accidentally it’s the best routine app I’ve ever used. I use clock periods instead of calendar periods to eliminate the incentive to start at the very beginning of the block (i.e. scheduling) and so that “overachievers” wouldn’t be waiting so long for the next calendar period that they forget the system, but this also has the side effect that routines tracked earlier the previous day will now have a stronger signal to do it earlier again the next and vice versa. With a 2hr default grace period, there is flexibility (++resilience: you don’t give up on your system the moment things don’t go to plan) at the same time as it encourages you generally to keep a pattern of habits that work for you at the same time as it encourages you generally to keep to a pattern of habits that work for you.

A rant on self-scheduling

Ideologically, I just despise self-scheduling. Spontaneity is in all things beautiful. Forgetfulness is just spontaneity in the negative, no less an exercise of freedom. Without scheduling we would have much less to forget, and for that I respect scheduling. However, I have zero respect for self-scheduling. “Sorry, I can’t [be a normal fun person],” says the self-scheduler, “I have to do this thing that nobody told me I had to and that I don’t even want to do myself.” Neither is there spontaneous beauty in forgetting self-orders–you are just back where you started except now also a failure. When I “succeed” in perfectly following my elaborate self-scheduling, it means that I accomplished something so mundane that I had already totally understood it before I even began.

Inspiration

The initial name was “toDoom” as the interface was inspired by The World’s Most Dangerous Writing App, and I intentionally hadn’t handled progress bar overflow, resulting in a crash and arbitrary data loss.

Can I have x feature to track something that I could just as easily track in the plain text

notes? no.

State management warning

I would HIGHLY recommend backing up your state file. State management is so far from stabilized, backwards compatibility is in no way guaranteed. As far as I’m concerned, your data may be lost at any time.

Modules

utils

Structs

Cli

Prac: a dead-simple practice-cultivating utility.

Config

Practice

State

Enums

GracePeriod 🔒

SubCommand 🔒

Functions

get_state_path 🔒

main 🔒