Enum prac::SubCommand

pub(crate) enum SubCommand {
    List {
        cumulative: bool,
        period: bool,
    },
    Add {
        name: String,
        notes: Option<String>,
        period: u64,
        time_unit: TimeUnit,
    },
    Track {
        name: Option<String>,
        time: u64,
        time_unit: TimeUnit,
        log: bool,
    },
    Log {
        name: Option<String>,
    },
    EditPeriod {
        name: Option<String>,
        period: u64,
        time_unit: TimeUnit,
    },
    Remove {
        name: Option<String>,
    },
    Rename {
        current_name: String,
        new_name: String,
    },
    Reset,
    StateLocation,
}

Variants

List

Fields

cumulative: bool

Show cumulative hours tracked alongside practices.

period: bool

Show period of practice alongside practices

List practices, with a progress bar showing time through period since last practice. For options, see subcommand –help

Add

Fields

name: String

A (unique) name for the practice.

notes: Option<String>

Some notes to outline goals and log progress. Opens $EDITOR if not provided.

period: u64

Anticipated time period between practice sessions. (There is a 2 hr grace period by default.)

time_unit: TimeUnit

Add a new practice.

Track

Fields

name: Option<String>

Specify, or leave blank to fuzzy search.

time: u64

How long you practiced for. This is added to the cumulative time displayed in prac list --cumulative.

time_unit: TimeUnit

log: bool

An optional shortcut to prac log when you’re done.

Mark a practice completion, logging time.

Log

Fields

name: Option<String>

Specify practice to edit, or leave blank to fuzzy search.

Edit practice notes w/ your $EDITOR. Each practice has its own log.

EditPeriod

Fields

name: Option<String>

Specify, or leave blank to fuzzy search.

period: u64

Anticipated time period between practice sessions. (There is a 2 hr grace period by default.)

time_unit: TimeUnit

Edit practice period.

Remove

Fields

name: Option<String>

Specify name of practice to remove, or leave blank to fuzzy search.

Delete a practice.

Rename

Fields

current_name: String

new_name: String

Rename practice.

Reset

Reset progress bars. Equivalent to tracking all practices w/ zero time.

StateLocation

Show state file location.

State is stored in $PRACTICE_PATH, $PRACTICE_HOME/prac.json, or $XDG_DATA_HOME/prac/prac.json or ~/.prac.json if none of previous are set.